- Adicionado novo modo de remover produtos.
obs: Lembrar de:
1) Executar o script que esta no pacote "scripts". (Obrigatório)
2) Executar o PopulaBanco. (Opcional)
3) Se houver erro no PopulaBanco repita o processo (normalmente até a terceira tentativa funciona). (Opcional)

+++ PDS-SGAP/src/tela/locador/meusprodutos/ProdutoDetalhesTela.java
@@ -1,5 +1,6 @@
 package tela.locador.meusprodutos;
 
+import dao.AluguelDAO;
 import modelo.painel.PainelModeloImagens;
 import modelo.lista.ListaModeloCategorias;
 import dao.ProdutoDAO;
@@ -335,12 +336,18 @@ public class ProdutoDetalhesTela extends javax.swing.JFrame {
     }//GEN-LAST:event_bAlterarActionPerformed
 
     private void bExcluirActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bExcluirActionPerformed
-        int resposta = JOptionPane.showConfirmDialog(null, "Deseja excluir o registro: " + produto.getNome() + "?", "Excluir Produto", JOptionPane.YES_NO_OPTION);
-        if(resposta == JOptionPane.YES_OPTION){
-            JOptionPane.showMessageDialog(null, "Produto excluido com sucesso", "Produto Excluido", JOptionPane.INFORMATION_MESSAGE);
-            dispose();
-            ProdutoDAO produtoDAO = new ProdutoDAO();
-            produtoDAO.removeProduto(produto);
+        AluguelDAO aluguelDAO = new AluguelDAO();
+        boolean produtoAlugado = aluguelDAO.isAlugado(produto);
+        if (produtoAlugado) {
+            JOptionPane.showMessageDialog(null, "Produto não pode ser excluido, pois esta sendo alugado no momento", "Aviso", JOptionPane.INFORMATION_MESSAGE);
+        } else {
+            int resposta = JOptionPane.showConfirmDialog(null, "Deseja excluir o registro: " + produto.getNome() + "?", "Excluir Produto", JOptionPane.YES_NO_OPTION);
+            if(resposta == JOptionPane.YES_OPTION){
+                ProdutoDAO produtoDAO = new ProdutoDAO();
+                produtoDAO.removeProduto(produto);
+                JOptionPane.showMessageDialog(null, "Produto excluido com sucesso", "Produto Excluido", JOptionPane.INFORMATION_MESSAGE);
+                dispose();
+            }
         }
     }//GEN-LAST:event_bExcluirActionPerformed
